refactor: tier-2 architecture review bundle (#61)
* refactor(types): unify duplicate SlotRange, NodeRole, and OperationType

Three SlotRange structs collapse into one canonical type at
slots::distribution::SlotRange with u16 fields (the correct domain
type — slots are 0..=16383). Removed duplicates from client::types
and controller::cluster_topology, which now re-export the canonical
type. Added SlotRange::new, ::single, ::parse_valkey, ::count,
::contains, ::iter helpers and the inevitable i32::from boundary
conversions where K8s/CRD types need i32.

Two NodeRole enums collapse into client::types::NodeRole with three
variants (Master, Replica, Unknown). Display divergence is now
explicit: as_valkey_str() for the wire format ("master"/"slave"/"?")
and as_display_str() for human text ("master"/"replica"/"unknown")
rather than hidden in two parallel Display impls. cluster_topology
re-exports for backward compat.

Two OperationType enums collapse into
controller::operation_coordination::OperationType, now Copy and
FromStr. operation_lock re-exports.

Removes 8 newly-redundant `as u16` casts and the i32 SlotRange
test helper signature.


* refactor(resources): split oversized helpers in statefulset.rs

build_valkey_extra_flags (124 lines) extracted into base_cluster_flags
plus four append_* helpers (replication, min_replicas, acl, tls), each
under 100 lines.

generate_metrics_exporter_container (130 lines) extracted into
build_exporter_env_vars, build_exporter_resource_requirements, and
build_exporter_http_probe. The new probe builder is reusable across
the readiness/liveness/startup probes.

Behavior unchanged — existing test fixtures pass byte-identically.


* refactor(controller): split cluster_reconciler, add tracing, classify retries

cluster_reconciler.rs (was 3433 lines) split into focused submodules:
  controller/cluster_reconciler/scaling.rs   — execute_scale_down and
                                                friends, plus 16 unit tests
  controller/cluster_reconciler/deletion.rs  — handle_deletion,
                                                forget_nodes_with_quorum
  controller/cluster_reconciler/status.rs    — update_status,
                                                ClusterHealthStatus,
                                                StatusUpdate<'a>
reconcile() is now ~120 lines with a dispatch_phase() helper for the
phase match. Submodule items use pub(crate) only where called from
the parent module.

StatusUpdate<'a> struct with .apply() replaces the 7-positional-param
update_status(); all 5 call sites updated to construct the struct.

#[tracing::instrument] added on reconcile entry points, all 11 phase
handlers, and key submodule operations (20 attributes total). skip(...)
prevents logging large structs; structured fields capture name,
namespace, phase.

is_retryable() refined: ValkeyError variants are now classified
per-variant instead of blanket-true. Connection / Timeout /
ClusterNotReady / IO / fred-cluster errors retry; Parse /
InvalidConfig / SlotMigrationFailed do not.

PatchParams::apply()+Patch::Merge misnomers in scope (common.rs,
cluster_phases.rs, cluster_reconciler.rs, status.rs,
upgrade_reconciler.rs) corrected. operation_lock.rs and
operation_coordination.rs left for the parallel-worktree integration.


* refactor(operations): correct PatchParams::apply()+Patch::Merge mismatches

Three remaining sites in operation_coordination.rs and operation_lock.rs
that were out of scope for the parallel reconciler refactor: they used
PatchParams::apply(FIELD_MANAGER) with Patch::Merge, but apply implies
SSA. Switched to PatchParams::default() to match the established
non-SSA pattern. Dropped the now-unused FIELD_MANAGER import in
operation_coordination.rs.


* style: rustfmt — collapse short function calls


---------

+++ src/controller/cluster_init.rs
@@ -479,7 +479,7 @@ pub async fn assign_slots_to_masters(
     for master in cluster_nodes.masters() {
         for slot_range in &master.slots {
             for slot in slot_range.start..=slot_range.end {
-                assigned_slots.insert(slot as u16);
+                assigned_slots.insert(slot);
             }
         }
     }
